pkg/sysinfo: stub out parsing cpusets on non-linux
