+++ pkg/sysinfo/sysinfo_other.go
@@ -6,3 +6,7 @@ package sysinfo // import "github.com/docker/docker/pkg/sysinfo"
 func New(options ...Opt) *SysInfo {
 	return &SysInfo{}
 }
+
+func isCpusetListAvailable(string, string) (bool, error) {
+	return false, nil
+}
